+++ R/jamma.r
@@ -124,7 +124,7 @@ NULL
 #'    to substitute as the first color from `colramp`. This method keeps the
 #'    color ramp consistent, but changes the background color to highlight
 #'    the plot panel as an outlier.
-#' @param applyRangeCeiling see \code{\link{plotSmoothScatter}} for details,
+#' @param applyRangeCeiling see \code{\link{jamba::plotSmoothScatter}} for details,
 #'    logical whether to apply a noise floor and ceiling to display points
 #'    outside the plot region at the boundaries of the plot. This parameter
 #'    is useful to set FALSE when cropping a plot, in order to remove outlier
@@ -145,10 +145,10 @@ NULL
 #'    of each plot panel.
 #' @param titleBoxColor vector of colors applied to title text. When doTitleBox=TRUE
 #'    one or no value is supplied, it defines colors using
-#'    \code{\link{setTextContrastColor}} to use a contrasting color.
+#'    \code{\link{jamba::setTextContrastColor}} to use a contrasting color.
 #' @param titleColor vector of colors applied to title text. When doTitleBox=TRUE
 #'    one or no value is supplied, it defines colors using
-#'    \code{\link{setTextContrastColor}} to use a contrasting color.
+#'    \code{\link{jamba::setTextContrastColor}} to use a contrasting color.
 #' @param doTitleBox logical whether to draw plot titles using a colored box.
 #' @param titleFont integer font compatible with \code{par("font")}. Values
 #'    are recycled across panels.
@@ -179,11 +179,11 @@ NULL
 #'    alpha transparency to use for the polygon color, based upon
 #'    highlightColor.
 #' @param smoothPtCol color used when nrpoints>0, and
-#'    \code{\link{plotSmoothScatter}} draws this many points in the extremities.
+#'    \code{\link{jamba::plotSmoothScatter}} draws this many points in the extremities.
 #' @param margins vector of margins compatible with \code{par("mar")}. Defaults
 #'    are applied, but provided here for convenient override.
 #' @param useRaster logical whether to define \code{useRaster=TRUE} in
-#'    \code{\link{plotSmoothScatter}}, which ultimately calls
+#'    \code{\link{jamba::plotSmoothScatter}}, which ultimately calls
 #'    \code{\link{imageDefault}}. When TRUE, it creates a much smaller
 #'    plot object, because it returns a raster image instead of a set of
 #'    rectangles when rendering the image.
@@ -242,14 +242,14 @@ NULL
 #'    zero. For some platform data , it might be useful to define a
 #'    \code{filterFloor} roughly equivalent to its noise threshold.
 #' @param transFactor numeric adjustment used by
-#'    \code{\link{plotSmoothScatter}}. The default is defined, but available
+#'    \code{\link{jamba::plotSmoothScatter}}. The default is defined, but available
 #'    here as a convenient override. The value is used in a power function in
 #'    the form \code{transformation=function(x)x^transFactor}. Lower values
 #'    make the scatterplot point density more intense.
 #' @param nrpoints integer or NULL, the number of points to display on the
-#'    extremity, as used by \code{\link{plotSmoothScatter}}.
+#'    extremity, as used by \code{\link{jamba::plotSmoothScatter}}.
 #' @param smoothScatterFunc function used to plot smooth scatter plot, by
-#'    default \code{\link{plotSmoothScatter}}.
+#'    default \code{\link{jamba::plotSmoothScatter}}.
 #' @param ablineH,ablineV numeric vector indicating where to draw indicator
 #'    lines, at these horizontal or vertical positions, respectively.
 #' @param doTxtplot logical whether to plot results in colored text output,
@@ -285,14 +285,14 @@ NULL
 #'    a sample outlier would need to have greater than five times median
 #'    absolute deviation compared to its center group, based only upon
 #'    measurements above noise.
-#' @param fillBackground logical sent to \code{\link{plotSmoothScatter}}
+#' @param fillBackground logical sent to \code{\link{jamba::plotSmoothScatter}}
 #'    indicating whether to fill the plot panel with the background color,
 #'    which is typically the first value in the \code{colramp}. Set to TRUE
 #'    when axis ranges may be defined which are outside the range of values
 #'    for any particular column in \code{object}.
 #' @param verbose logical indicating whether to print verbose output.
 #' @param ... additional parameters sent to downstream functions,
-#'    \code{\link{plotSmoothScatter}}, \code{\link{centerGeneData}}.
+#'    \code{\link{jamba::plotSmoothScatter}}, \code{\link{centerGeneData}}.
 #'
 #' @return List of data used by jamma() to produce an MA-plot, sufficient
 #'          to use as input to produce another MA-plot.
@@ -355,7 +355,7 @@ jammaplot <- function
  filterFloorReplacement=filterFloor,
  transFactor=0.18,
  nrpoints=50,
- smoothScatterFunc=plotSmoothScatter,
+ smoothScatterFunc=jamba::plotSmoothScatter,
  applyRangeCeiling=TRUE,
  doTxtplot=FALSE,
  ablineV=0, ablineH=c(-2,0,2),
@@ -455,7 +455,7 @@ jammaplot <- function
    }
    firstColor <- head(colramp(10), 1);
    if (length(colrampOutlier) > 0) {
-      if (igrepHas("character", class(colrampOutlier))) {
+      if (jamba::igrepHas("character", class(colrampOutlier))) {
          colrampOutlier <- colorRampPalette(getColorRamp(colrampOutlier));
       }
    }
@@ -502,14 +502,14 @@ jammaplot <- function
    }
    titleCex <- rep(titleCex, length.out=nsamples);
    if (verbose) {
-      printDebug("jammaplot(): ",
+      jamba::printDebug("jammaplot(): ",
          "titleCex:", head(titleCex, 10), ", length=", length(titleCex));
    }
    titleBoxColor <- rep(titleBoxColor, length.out=nsamples);
    ## Adjust titleColor to have contrast from the titleBoxColor
    if (length(titleColor) <= 1) {
       if (doTitleBox) {
-         titleColor <- setTextContrastColor(titleBoxColor);
+         titleColor <- jamba::setTextContrastColor(titleBoxColor);
       } else {
          if (length(titleColor) == 0) {
             titleColor <- "black";
@@ -547,7 +547,7 @@ jammaplot <- function
       mvaDatas <- object;
    } else {
       if (verbose) {
-         printDebug("jammaplot(): ",
+         jamba::printDebug("jammaplot(): ",
             "Processing matrix input type, dim(object)", dim(object));
       }
       if (filterNA) {
@@ -603,9 +603,9 @@ jammaplot <- function
          }
       }
       if (verbose) {
-         printDebug("jammaplot(): ",
+         jamba::printDebug("jammaplot(): ",
             "Processing matrix input type, dim(object)", dim(object));
-         printDebug("jammaplot(): ",
+         jamba::printDebug("jammaplot(): ",
             "Processing matrix input type, dim(y):", dim(y));
       }
       if (is.null(xlim)) {
@@ -627,10 +627,10 @@ jammaplot <- function
          controlSamples=controlSamples,
          verbose=verbose);
       if (verbose) {
-         printDebug("jammaplot(): ",
+         jamba::printDebug("jammaplot(): ",
             "objectCtr:");
-         ch(objectCtr, rows=2);
-         printDebug("head(y): ", round(head(y), digits=2),
+         print(head(objectCtr));
+         jamba::printDebug("head(y): ", round(head(y), digits=2),
             fgText=c("orange", "lightblue"));
       }
    }
@@ -639,7 +639,7 @@ jammaplot <- function
    ## so we can test for outliers beforehand
    if (!gaveMVA) {
       if (verbose) {
-         printDebug("jammaplot(): ",
+         jamba::printDebug("jammaplot(): ",
             "Calculating MVA data.");
       }
       for (i in whichSamples) {
@@ -653,7 +653,7 @@ jammaplot <- function
             #x <- objectCtr[,i] + y;
             x <- objectCtr[,i] + yUse;
             if (verbose) {
-               printDebug("jammaplot(): ",
+               jamba::printDebug("jammaplot(): ",
                   "Applying unique x per centerGroups");
             }
          } else {
@@ -682,12 +682,12 @@ jammaplot <- function
    ## we are comparing with zero, and we are not subtracting an actual median value
    if (length(outlierMAD) > 0) {
       if (verbose) {
-         printDebug("Calculating MVA data MAD factors.");
+         jamba::printDebug("Calculating MVA data MAD factors.");
       }
       mvaMADs <- sapply(nameVector(whichSamples, colnames(object)[whichSamples]),
          function(i){
          if (verbose) {
-            printDebug("   i:", i, c("orange", "lightblue"));
+            jamba::printDebug("   i:", i, c("orange", "lightblue"));
          }
          mvaData <- mvaDatas[[i]];
          median(abs(mvaData[abs(mvaData[,"x"]) >= outlierRowMin,"y",drop=FALSE]),
@@ -712,11 +712,11 @@ jammaplot <- function
       names(mvaMADs) <- whichSamples;
       mvaMADoutliers <- colnames(object)[whichSamples][which(mvaMADfactors >= outlierMAD)];
       if (verbose) {
-         printDebug("mvaMADs:");
+         jamba::printDebug("mvaMADs:");
          print(format(digits=2, mvaMADs));
-         printDebug("mvaMADfactors:");
+         jamba::printDebug("mvaMADfactors:");
          print(format(digits=2, mvaMADfactors));
-         printDebug("mvaMADoutliers:", mvaMADoutliers);
+         jamba::printDebug("mvaMADoutliers:", mvaMADoutliers);
       }
       attr(mvaDatas, "mvaMADs") <- mvaMADs;
       attr(mvaDatas, "mvaMADoutliers") <- mvaMADoutliers;
@@ -735,19 +735,19 @@ jammaplot <- function
          mvaData <- mvaDatas[[i]];
          iPanelNumber <- iPanelNumber + 1;
          if (verbose) {
-            printDebug("iPanelNumber: ", iPanelNumber,
+            jamba::printDebug("iPanelNumber: ", iPanelNumber,
                fgText=c("orange", "lightgreen"));
          }
          if (length(blankPlotPos) > 0) {
             while(iPanelNumber %in% blankPlotPos) {
                nullPlot(doBoxes=FALSE);
                if (verbose) {
-                  printDebug("   Inserted a blank panel at position: ",
+                  jamba::printDebug("   Inserted a blank panel at position: ",
                      iPanelNumber, fgText=c("orange", "lightblue"));
                }
                iPanelNumber <- iPanelNumber + 1;
                if (verbose) {
-                  printDebug("new iPanelNumber: ", iPanelNumber,
+                  jamba::printDebug("new iPanelNumber: ", iPanelNumber,
                      fgText=c("orange", "lightgreen"));
                }
             }
@@ -836,11 +836,13 @@ jammaplot <- function
                      y=yValues),
                      returnClass="matrix");
                   if (is.null(highlightPolygonAlpha)) {
-                     highlightPolyAlpha <- col2alpha(highlightColor[[highI]]) / 2;
+                     highlightPolyAlpha <- jamba::col2alpha(highlightColor[[highI]]) / 2;
                   }
-                  highlightPolyColor <- alpha2col(highlightColor[[highI]],
+                  highlightPolyColor <- jamba::alpha2col(highlightColor[[highI]],
                      alpha=highlightPolyAlpha);
-                  polygon(hiHull, add=TRUE, col=highlightPolyColor,
+                  polygon(hiHull,
+                     add=TRUE,
+                     col=highlightPolyColor,
                      border=highlightPolyColor);
                }
                ## Draw the highlighted points
@@ -848,7 +850,7 @@ jammaplot <- function
                   y=yValues,
                   pch=highlightPch[[highI]],
                   bg=highlightColor[[highI]],
-                  col=makeColorDarker(darkFactor=1.2,
+                  col=jamba::makeColorDarker(darkFactor=1.2,
                      sFactor=1.5,
                      highlightColor[[highI]]),
                   xpd=FALSE,
@@ -857,50 +859,79 @@ jammaplot <- function
          }
          ## Optionally draw title boxes as labels per plot
          if (verbose) {
-            printDebug("jammaplot(): ",
+            jamba::printDebug("jammaplot(): ",
                "titleFont:",
                titleFont);
          }
-         titleBoxTextColor <- ifelse(colnames(object)[i] %in% mvaMADoutliers,
-            blendColors(rep(c(titleColor[i], "red"), c(3,1))),
-            titleColor[i]);
+         titleBoxTextColor <- titleColor[i];
+         #titleBoxTextColor <- ifelse(colnames(object)[i] %in% mvaMADoutliers,
+         #   blendColors(rep(c(titleColor[i], "red"), c(3,1))),
+         #   titleColor[i]);
          if (doTitleBox) {
             if (verbose) {
-               printDebug("jammaplot(): ",
-                  "   titleBoxColor:", titleBoxColor[i],
-                  list("orange",titleBoxColor[i]));
+               jamba::printDebug("jammaplot(): ",
+                  "   titleBoxColor:",
+                  titleBoxColor[i],
+                  bgText=list(NA, titleBoxColor[i]),
+                  fgText=list("orange", titleColor[i]));
             }
             parXpd <- par("xpd");
             par("xpd"=NA);
-            legend("topright",
-               inset=0.02,
-               cex=titleCex[i],
-               text.font=titleFont[i],
-               legend=paste(titleText, groupSuffix[i]),
-               adj=c(0,0.5),
-               bg=titleBoxColor[i],
-               title.col=titleBoxTextColor,
-               text.col=titleBoxTextColor,
-               title=maintitle);
-            ## Subtitle using tricks to center the label
-            if (!is.null(subtitle)) {
-               b1 <- legend("bottom", inset=0.02, plot=FALSE,
+            if (1 == 2) {
+               legend("topright",
+                  inset=0.02,
                   cex=titleCex[i],
                   text.font=titleFont[i],
-                  legend=subtitle,
-                  title=NULL);
-               rect(xleft=b1$rect$left,
-                  ybottom=b1$rect$top-b1$rect$h,
-                  xright=b1$rect$left+b1$rect$w,
-                  ytop=b1$rect$top,
-                  col=titleBoxColor[i],
-                  border="black");
-               text(x=b1$rect$left+b1$rect$w/2,
-                  y=b1$rect$top-b1$rect$h/2,
-                  labels=subtitle,
-                  cex=titleCex[i],
-                  font=titleFont[i],
-                  col=titleBoxTextColor);
+                  legend=paste(titleText, groupSuffix[i]),
+                  adj=c(0,0.5),
+                  bg=titleBoxColor[i],
+                  title.col=titleBoxTextColor,
+                  text.col=titleBoxTextColor,
+                  title=maintitle);
+            } else {
+               ## Use drawLabels() for more control over the text box
+               drawLabels(preset="top",
+                  txt=paste(titleText, groupSuffix[i]),
+                  boxColor=titleBoxColor[i],
+                  boxBorderColor=jamba::makeColorDarker(titleBoxColor[i]),
+                  labelCol=titleColor[i],
+                  labelCex=titleCex[i],
+                  drawBox=doTitleBox,
+                  boxCexAdjust=c(1.1,1.3),
+                  font=titleFont[i]);
+            }
+            ## Subtitle using tricks to center the label
+            if (length(subtitle) > 0) {
+               if (1 == 2) {
+                  b1 <- legend("bottom", inset=0.02, plot=FALSE,
+                     cex=titleCex[i],
+                     text.font=titleFont[i],
+                     legend=subtitle,
+                     title=NULL);
+                  rect(xleft=b1$rect$left,
+                     ybottom=b1$rect$top-b1$rect$h,
+                     xright=b1$rect$left+b1$rect$w,
+                     ytop=b1$rect$top,
+                     col=titleBoxColor[i],
+                     border="black");
+                  text(x=b1$rect$left+b1$rect$w/2,
+                     y=b1$rect$top-b1$rect$h/2,
+                     labels=subtitle,
+                     cex=titleCex[i],
+                     font=titleFont[i],
+                     col=titleBoxTextColor);
+               } else {
+                  ## Use drawLabels() for more control over the text box
+                  drawLabels(preset="bottom",
+                     txt=subtitle[i],
+                     boxColor=titleBoxColor[i],
+                     boxBorderColor=jamba::makeColorDarker(titleBoxColor[i]),
+                     labelCol=titleColor[i],
+                     labelCex=titleCex[i],
+                     drawBox=doTitleBox,
+                     boxCexAdjust=c(1.1,1.3),
+                     font=titleFont[i]);
+               }
             }
             par("xpd"=parXpd);
          } else {
@@ -948,13 +979,13 @@ jammaplot <- function
       if (length(blankPlotPos) > 0) {
          iPanelNumber <- iPanelNumber + 1;
          if (verbose) {
-            printDebug("iPanelNumber: ", iPanelNumber,
+            jamba::printDebug("iPanelNumber: ", iPanelNumber,
                fgText=c("orange", "lightgreen"));
          }
          while(iPanelNumber %in% blankPlotPos) {
             nullPlot(doBoxes=FALSE);
             if (verbose) {
-               printDebug("Inserted a blank panel at position: ", iPanelNumber,
+               jamba::printDebug("Inserted a blank panel at position: ", iPanelNumber,
                   fgText=c("orange", "lightblue"));
             }
             iPanelNumber <- iPanelNumber + 1;
@@ -1110,7 +1141,7 @@ centerGeneData <- function
    }
    if (needsLog) {
       if (verbose) {
-         printDebug("centerGeneData(): ",
+         jamba::printDebug("centerGeneData(): ",
             "applying log2 transform:",
             "log(1 + x)");
       }
@@ -1123,9 +1154,9 @@ centerGeneData <- function
       centerControl=colnames(indata) %in% controlCols)));
    #centerDF <- mmixedOrderDF(centerDF, byCols=c(2,-3));
    if (verbose) {
-      printDebug("centerGeneData(): ",
+      jamba::printDebug("centerGeneData(): ",
          "dim(indata)", dim(indata));
-      printDebug("centerGeneData(): ",
+      jamba::printDebug("centerGeneData(): ",
          "dim(centerDF)", dim(centerDF));
       print(head(centerDF));
    }
@@ -1140,7 +1171,7 @@ centerGeneData <- function
 
    ## Optionally center groups separately
    if (length(centerGroups) > 0) {
-      if (!igrepHas("list", class(centerGroups))) {
+      if (!jamba::igrepHas("list", class(centerGroups))) {
          if (length(names(centerGroups)) > 0) {
             if (all(colnames(indata) %in% names(centerGroups))) {
                centerGroups <- centerGroups[colnames(indata)];
@@ -1270,7 +1301,7 @@ centerGeneData <- function
       }
    }
    if (hasCharColumns) {
-      printDebug("Keeping non-numeric columns as-is.");
+      jamba::printDebug("Keeping non-numeric columns as-is.");
       centeredData <- cbind(indataChar, centeredData)
    }
    attr(centeredData, "centerDF") <- centerDF;
@@ -1328,25 +1359,25 @@ points2polygonHull <- function
    if (!suppressPackageStartupMessages(require(rgeos))) {
       stop("points2polygonHull() requires the rgeos package gConvexHull().");
    }
-   if (igrepHas("matrix|data.*frame|tbl", class(x))) {
+   if (jamba::igrepHas("matrix|data.*frame|tbl", class(x))) {
       if (ncol(x) != 2) {
          stop(paste0("points2polygonHull() requires x as a",
             " matrix,data.frame,tbl",
             " to contain 2 columns."));
       }
       x <- SpatialPoints(x);
-   } else if (igrepHas("SpatialPoints", class(x))) {
+   } else if (jamba::igrepHas("SpatialPoints", class(x))) {
       #SpatialPoints(x)
    }
 
    ## Determine convex hull coordinates
    gH <- gConvexHull(x);
    if (returnClass %in% "matrix") {
-      if (igrepHas("polygons", class(gH))) {
+      if (jamba::igrepHas("polygons", class(gH))) {
          gHcoords <- gH@polygons[[1]]@Polygons[[1]]@coords;
-      } else if (igrepHas("lines", class(gH))) {
+      } else if (jamba::igrepHas("lines", class(gH))) {
          gHcoords <- gH@lines[[1]]@Lines[[1]]@coords;
-      } else if (igrepHas("points", class(gH))) {
+      } else if (jamba::igrepHas("points", class(gH))) {
          gHcoords <- gH@coords;
       }
    } else {
@@ -1355,4 +1386,602 @@ points2polygonHull <- function
    gHcoords;
 }
 
+#' Draw text labels on a base R plot
+#'
+#' Draw text labels on a base R plot
+#'
+#' This function takes a vector of coordinates and text labels,
+#' and draws the labels with colored rectangles around each label
+#' on the plot. Each label can have unique font, cex, and color,
+#' and are drawn using vectorized operations.
+#'
+#' TODO: In future allow rotated text labels. Not that useful within
+#' a plot panel, but sometimes useful when draw outside a plot, for
+#' example axis labels.
+#'
+#' @return invisible data.frame containing label coordinates used
+#' to draw labels. This data.frame can be manipulated and provided
+#' as input to `drawLabels()` for subsequent customized label
+#' positioning.
+#'
+#' @family jam plot functions
+#'
+#' @param newCoords optional `data.frame` typically as a result of
+#'    a previous call to `drawLabels()`. In general, it should contain
+#'    colnames equivalent to the function parameters of `drawLabels()`.
+#' @param x,y vector of x- and y- coordinates.
+#' @param txt character vector of labels, length equal to `x` and `y`.
+#' @param lx,ly optional vector of segment endpoint coordinates, used
+#'    to draw a line from x,y coordinates to the segment lx,ly coordinate.
+#' @param segmentLwd,segmentCol vector of segment line widths, and colors,
+#'    respectively. Each vector will be recycled to `length(txt)` as
+#'    needed.
+#' @param drawSegments logical whether to draw segments, where applicable.
+#' @param boxBorderColor vector of colors used for the box border around
+#'    each label.
+#' @param boxColor vector of colors used for the box background behind
+#'    each label.
+#' @param boxLwd vector of box line widths, sent to `graphics::rect()`,
+#'    this vector will be recycled to `length(txt)`.
+#' @param drawBox logical whether to draw boxes behind each text label.
+#' @param drawLabels logical whether to draw each text label.
+#' @param font vector of font values as described in `graphics::par()`,
+#'    where 1=normal, 2=bold, 3=italics, 4=bold-italics.
+#' @param labelCex vector of cex values used for text labels. This vector
+#'    will be recycled to `length(txt)` as needed.
+#' @param boxCexAdjust numeric vector length=2, used to expand the x-width
+#'    and y-height of the box around around text labels.
+#' @param labelCol vector of label colors, by default it calls
+#'    `jamba::setTextContrastColor()` to generate a color to contrast
+#'    the background box color.
+#' @param doPlot logical whether to perform any plot operations. Set
+#'    `FALSE` to calculate coordinates and return a `data.frame` of
+#'    label coordinates, which can then be manipulated before calling
+#'    `drawLabels()` again.
+#' @param xpd value compatible with `par("xpd")`, where NA allows labels
+#'    anywhere in the device region, TRUE retricts labels within the figure
+#'    region, and FALSE restricts labels within the plot region.
+#' @param preset vector of text values used to position text labels relative
+#'    to the x,y coordinate, where "topleft" will position the label so the
+#'    entire label box is top-left of the point, therefore the point will be
+#'    at the bottom-right corner of the label box. When `preset` is anything
+#'    by `"none"` the `adjX` and `adjY` values are ignored.
+#' @param adjX,adjY the text adjustment of labels relative to the x,y
+#'    coordinate. The values are recycled to `length(txt)`.
+#' @param trimReturns logical whether to trim leading and trailing return
+#'    (newline) characters from labels.
+#' @param verbose logical whether to print verbose output.
+#' @param ... additional arguments are passed to `graphics::segments()` when
+#'    segments are drawn, to `graphics::rect()` when label boxes are drawn,
+#'    and to `graphics::text()` when text labels are drawn.
+#'
+#' @examples
+#' jamba::nullPlot();
+#' drawLabels(x=par("usr")[1],
+#'    y=par("usr")[4],
+#'    txt="Top-left\nof plot",
+#'    preset="bottomright",
+#'    boxColor="blue4");
+#'
+#' drawLabels(x=par("usr")[2],
+#'    y=par("usr")[3],
+#'    txt="Bottom-right\nof plot",
+#'    preset="topleft",
+#'    boxColor="green4");
+#'
+#' drawLabels(x=mean(par("usr")[1:2]),
+#'    y=mean(par("usr")[3:4]),
+#'    txt="Center\nof plot",
+#'    preset="center",
+#'    boxColor="purple");
+#'
+#' points(x=c(par("usr")[1], par("usr")[2],
+#'       mean(par("usr")[1:2])),
+#'    y=c(par("usr")[4], par("usr")[3],
+#'       mean(par("usr")[3:4])),
+#'    pch=20,
+#'    col="red",
+#'    xpd=NA);
+#'
+#' @export
+drawLabels <- function
+(newCoords=NULL,
+ x=NULL,
+ y=NULL,
+ txt=NULL,
+ lx=NULL,
+ ly=NULL,
+ segmentLwd=1,
+ segmentCol="#00000088",
+ drawSegments=TRUE,
+ boxBorderColor="#000000AA",
+ boxColor="#DDAA77",
+ boxLwd=1,
+ drawBox=TRUE,
+ drawLabels=TRUE,
+ font=1,
+ labelCex=0.8,
+ boxCexAdjust=c(1,1.2),
+ labelCol=jamba::alpha2col(alpha=0.8, jamba::setTextContrastColor(boxColor)),
+ doPlot=TRUE,
+ xpd=NA,
+ preset=c("default",
+   "top",
+   "topright",
+   "right",
+   "bottomright",
+   "topleft",
+   "left",
+   "bottomleft",
+   "bottom",
+   "center"),
+ adjPreset=c("default",
+   "top",
+   "topright",
+   "right",
+   "bottomright",
+   "topleft",
+   "left",
+   "bottomleft",
+   "bottom",
+   "center"),
+ adjX=0.5,
+ adjY=0.5,
+ trimReturns=TRUE,
+ verbose=FALSE,
+ ...)
+{
+   ## Purpose is to wrapper only the last portion of addNonOverlappingLabels()
+   ## which draws the labels, boxes, and segments after positions are determined
+   ## by addNonOverlappingLabels().
+   if (length(boxCexAdjust) == 0) {
+      boxCexAdjust <- 1;
+   }
+   boxCexAdjust <- rep(boxCexAdjust, length.out=2);
+   preset <- match.arg(preset);
+   adjPreset <- match.arg(adjPreset);
 
+
+   if (length(newCoords) == 0) {
+      ## Create a basic data.frame
+      if (jamba::igrepHas("top|bottom|left|right|center", c(preset, adjPreset))) {
+         if (verbose) {
+            jamba::printDebug("drawLabels(): ",
+               "Processing non-default preset and adjPreset values.");
+         }
+         presetL <- coordPresets(
+            preset=preset,
+            x=x,
+            y=y,
+            adjPreset=adjPreset,
+            adjX=adjX,
+            adjY=adjY);
+         x <- presetL$x;
+         y <- presetL$y;
+         adjX <- presetL$adjX;
+         adjY <- presetL$adjY;
+         preset <- presetL$preset;
+         adjPreset <- presetL$adjPreset;
+      }
+      newCoords <- data.frame(
+         x=x,
+         y=y,
+         txt=txt,
+         w=strwidth(txt,
+            font=font,
+            cex=labelCex),
+         h=strheight(txt,
+            font=font,
+            cex=labelCex),
+         hNudge=0,
+         labelCex=labelCex,
+         adjX=adjX,
+         adjY=adjY,
+         preset=preset,
+         adjPreset=adjPreset);
+      if (length(lx) > 0 && length(ly) > 0) {
+         newCoords$lx <- rep(lx, length.out=nrow(newCoords));
+         newCoords$ly <- rep(ly, length.out=nrow(newCoords));
+      } else {
+         drawSegments <- FALSE;
+      }
+
+      ## Make additional adjustments based upon leading or trailing "\n"
+      if (trimReturns) {
+         ## First determine the height of the buffer between two lines of text
+         labelCexU <- unique(newCoords$labelCex);
+         labelCexUh <- sapply(labelCexU, function(iCex){
+            bufferH <- (strheight("|\n|", cex=iCex) -
+                  2*(strheight("|", cex=iCex)));
+         });
+         #labelCexUh[match(newCoords$labelCex, labelCexU)]
+         if (jamba::igrepHas("\n$", newCoords$txt)) {
+            if (verbose) {
+               jamba::printDebug("drawLabels(): ",
+                  "adjusting adjY for trailing return characters.");
+               print(newCoords);
+            }
+            while(jamba::igrepHas("\n$", newCoords$txt)) {
+               iAdj <- jamba::igrep("\n$", newCoords$txt);
+               adjY[iAdj] <- adjY[iAdj] - newCoords$adjY[iAdj];
+               newCoords$txt[iAdj] <- sub("\n$", "", newCoords$txt[iAdj]);
+            }
+         }
+         if (jamba::igrepHas("^\n", newCoords$txt)) {
+            if (verbose) {
+               jamba::printDebug("drawLabels(): ",
+                  "adjusting adjY for leading return characters.");
+            }
+            while(jamba::igrepHas("^\n", newCoords$txt)) {
+               iAdj <- jamba::igrep("^\n", newCoords$txt);
+               adjY[iAdj] <- adjY[iAdj] + newCoords$adjY[iAdj];
+               newCoords$txt[iAdj] <- sub("^\n", "", newCoords$txt[iAdj]);
+            }
+         }
+         newCoords$adjY <- adjY;
+         newCoords$h <- strheight(newCoords$txt, cex=labelCex);
+      } else {
+         newCoords$txt <- gsub("\n", "\n|", newCoords$txt);
+      }
+
+      ## Now widen the label box using boxCexAdjust
+      if (verbose) {
+         jamba::printDebug("drawLabels(): ",
+            "adjusting w,h using boxCexAdjust");
+         print(newCoords);
+      }
+      ## New strategy intended to keep the bottom-left edge fixed
+      newCoords$h <- newCoords$h * boxCexAdjust[2];
+      newCoords$w <- newCoords$w * boxCexAdjust[1];
+      newCoords$x <- newCoords$x - adjX * newCoords$w;
+      newCoords$y <- newCoords$y - adjY * newCoords$h;
+   } else {
+      preset <- newCoords$preset;
+      adjPreset <- newCoords$adjPreset;
+      if (jamba::igrepHas("top|bottom|left|right|center", c(preset, adjPreset))) {
+         if (verbose) {
+            jamba::printDebug("drawLabels(): ",
+               "Processing non-default preset and adjPreset values.");
+         }
+         presetL <- coordPresets(
+            preset=preset,
+            x=newCoords$x,
+            y=newCoords$y,
+            adjPreset=adjPreset,
+            adjX=newCoords$adjX,
+            adjY=newCoords$adjY);
+         x <- presetL$x;
+         y <- presetL$y;
+         adjX <- presetL$adjX;
+         adjY <- presetL$adjY;
+         preset <- presetL$preset;
+         adjPreset <- presetL$adjPreset;
+      }
+      newCoords$x <- x;
+      newCoords$y <- y;
+      newCoords$adjX <- adjX;
+      newCoords$adjY <- adjY;
+      newCoords$preset <- preset;
+      newCoords$adjPreset <- adjPreset;
+   }
+
+   ## Add height and width if not supplied
+   if (!"h" %in% names(newCoords)) {
+      newCoords$h <- strheight(txt, cex=labelCex*1.1);
+   }
+   if (!"w" %in% names(newCoords)) {
+      newCoords$w <- strwidth(txt, cex=labelCex*1.1);
+   }
+
+   x <- newCoords$ptX;
+   y <- newCoords$ptY;
+   if (!"labelCex" %in% names(newCoords)) {
+      if (verbose) {
+         jamba::printDebug("drawLabels(): ",
+            "Defining all fixedLabels=FALSE");
+      }
+      newCoords[,"fixedLabels"] <- FALSE;
+   }
+   if (!"fixedLabels" %in% names(newCoords)) {
+      newCoords[,"fixedLabels"] <- TRUE;
+   }
+   whichLabels <- which(!newCoords$fixedLabels > 1);
+   if (verbose) {
+      jamba::printDebug("drawLabels(): ",
+         "head(whichLabels):",
+         head(whichLabels));
+   }
+
+   segmentCol <- rep(segmentCol, length.out=nrow(newCoords));
+   segmentLwd <- rep(segmentLwd, length.out=nrow(newCoords));
+
+   if (!"labelCex" %in% names(newCoords)) {
+      newCoords$labelCex <- rep(labelCex,
+         length.out=nrow(newCoords));
+   }
+   if (!"labelCol" %in% names(newCoords)) {
+      newCoords$labelCol <- rep(labelCol,
+         length.out=nrow(newCoords));
+   }
+
+   if (!"boxColor" %in% names(newCoords)) {
+      newCoords$boxColor <- rep(boxColor,
+         length.out=nrow(newCoords));
+   }
+   boxColor <- newCoords$boxColor;
+   if (!"boxBorderColor" %in% names(newCoords)) {
+      newCoords$boxBorderColor <- rep(boxBorderColor,
+         length.out=nrow(newCoords));
+   }
+   boxBorderColor <- newCoords$boxBorderColor;
+
+   ## Draw segments first so the box and labels will be printed on top of them
+   if (doPlot &&
+         drawSegments &&
+         any(whichLabels) &&
+         all(c("lx","ly") %in% names(newCoords))) {
+      ## Determine closest corner or flat side for each segment
+      x0l <- newCoords$x[whichLabels];
+      x0m <- newCoords$x[whichLabels] + newCoords$w[whichLabels]/2;
+      x0r <- newCoords$x[whichLabels] + newCoords$w[whichLabels];
+      y0b <- newCoords$y[whichLabels];
+      y0m <- newCoords$y[whichLabels] + newCoords$h[whichLabels]/2;
+      y0t <- newCoords$y[whichLabels] + newCoords$h[whichLabels];
+      x1 <- newCoords$lx[whichLabels];
+      y1 <- newCoords$ly[whichLabels];
+      sx0 <- ifelse(x0r < x1, x0r,
+         ifelse(x0l > x1, x0l,
+            x1));
+      sy0 <- ifelse(y0t < y1, y0t,
+         ifelse(y0b > y1, y0b,
+            y1));
+      if (verbose) {
+         headN <- 10;
+         jamba::printDebug("drawLabels(): ",
+            "Drawing line segments.");
+         jamba::printDebug("sx0: ",
+            head(sx0, headN));
+         jamba::printDebug("x1: ",
+            head(x1, headN));
+         jamba::printDebug("sy0: ",
+            head(sy0, headN));
+         jamba::printDebug("y1: ",
+            head(y1, headN));
+         jamba::printDebug("segmentCol[whichLabels]: ",
+            head(c(segmentCol[whichLabels]), headN), c("orange", "lightblue"));
+         jamba::printDebug("segmentLwd[whichLabels]: ",
+            head(c(segmentLwd[whichLabels]), headN), c("orange", "lightblue"));
+      }
+      segments(x0=sx0,
+         y0=sy0,
+         x1=newCoords$lx[whichLabels],
+         y1=newCoords$ly[whichLabels],
+         col=segmentCol[whichLabels],
+         lwd=segmentLwd[whichLabels],
+         xpd=xpd,
+         ...);
+   } else {
+      if (verbose) {
+         jamba::printDebug("drawLabels(): ",
+            "Drawing no segments.");
+      }
+   }
+
+   ## Due to a small bug in add.labels in the snippets package, the box and text
+   ## labels are not properly adjusted when the adj(1,1) parameter is supplied,
+   ## causing the boxes to be slightly offset in the y-axis direction relative
+   ## to the text.
+   ## The code below calculates the text to be the exact center of each box,
+   ## then forces the text() method below to use adj=c(0.5,0.5) which centers
+   ## the text exactly at this coordinate. Looks much nicer than other
+   ## options.
+   boxX1 <- newCoords$x;
+   boxX2 <- boxX1 + newCoords$w;
+   boxY1 <- newCoords$y - newCoords$hNudge;
+   boxY2 <- boxY1 + newCoords$h;
+   textX <- (boxX1 + boxX2) / 2;
+   textY <- (boxY1 + boxY2) / 2;
+
+   ## Draw boxes before labels so the labels will be printed on top
+   if (drawBox && doPlot && any(whichLabels)) {
+      if (verbose) {
+         jamba::printDebug("drawLabels(): ",
+            "Drawing rectangles.");
+         jamba::printDebug("drawLabels(): ",
+            "First rectangle:",
+            c(boxX1[whichLabels][1], boxX2[whichLabels][1],
+               boxY1[whichLabels][1], boxY2[whichLabels][1]));
+      }
+      ## Only draw boxes where there are characters to be printed
+      rect(xleft=boxX1[whichLabels],
+         ybottom=boxY1[whichLabels],
+         xright=boxX2[whichLabels],
+         ytop=boxY2[whichLabels],
+         col=boxColor[whichLabels],
+         border=boxBorderColor[whichLabels],
+         lwd=boxLwd,
+         xpd=xpd,
+         ...);
+   }
+   if (drawLabels && doPlot && any(whichLabels)) {
+      if (verbose) {
+         jamba::printDebug("drawLabels(): ",
+            "Printing labels head(txt): ",
+            paste(head(txt), collapse=", "));
+      }
+      text(x=textX[whichLabels],
+         y=textY[whichLabels],
+         font=font,
+         labels=newCoords$txt[whichLabels],
+         col=newCoords$labelCol[whichLabels],
+         cex=newCoords$labelCex[whichLabels],
+         outer=TRUE,
+         adj=c(0.5,0.5),
+         xpd=xpd,
+         ...);
+   }
+   invisible(newCoords);
+}
+
+#' Process coordinate adjustment presets
+#'
+#' Process coordinate adjustment presets
+#'
+#' This function is intended to be a convenient way to define
+#' coordinates using preset terms like "topleft", "bottom", "center".
+#'
+#' Similarly, it is intended to help define corresponding text
+#' adjustments, using `adj` compatible with `graphics::text()`,
+#' using preset terms like "bottomright", "center".
+#'
+#' When `preset` is `"default"`, the original `x,y` coordinates
+#' are used. Otherwise the `x,y` coordinates are defined using the
+#' plot region coordinates, where `"left"` uses `par("usr")[1]`,
+#' and `"top"` uses `par("usr")[4]`.
+#'
+#' When `adjPreset` is `"default"` it will use the `preset` to
+#' define a reciprocal text placement. For example when `preset="topright"`
+#' the text placement will be equivalent to `adjPreset="bottomleft"`.
+#' The `adjPreset` terms `"top"`, `"bottom"`, `"right"`, `"left"`,
+#' and `"center"` refer to the text label placement relative to
+#' `x,y` coordinate.
+#'
+#' If both `preset="default"` and `adjPreset="default"` the original
+#' `adjX,adjY` values are returned.
+#'
+#' The function is vectorized, and uses the longest input argument,
+#' so one can supply a vector of `preset` and it will return coordinates
+#' and adjustments of length equal to the input `preset` vector.
+#' The `preset` value takes priority over the supplied `x,y` coordinates.
+#'
+#' @return data.frame with colnames `x,y,adjX,adjY,preset,adjPreset`
+#' after adjustment, where the number of rows is determined by the
+#' longest input argument.
+#'
+#' @param preset character vector of coordinate positions, or "default"
+#'    to use the `x,y` coordinates.
+#' @param x,y numeric vectors indicating the default coordinates `x,y`.
+#' @param adjPreset character vector of text label positions, or
+#'    "default" to use `preset`, or when `preset="default"` the
+#'    `adjX,adjY` values are used.
+#' @param adjX,adjY numeric vectors indicating default text adjustment
+#'    values, as described for `adj` in `graphics::text()`.
+#' @param adjOffsetX,adjOffsetY numeric vector used to apply an offset
+#'    value to the `adjX,adjY` values, where positive values would
+#'    place a label farther away from center. Note these units are
+#'    relative to the text label size, when used with `graphics::text()`,
+#'    larger labels will be adjusted more than smaller labels.
+#' @param verbose logical indicating whether to print verbose output.
+#' @param ... additional arguments are ignored.
+#'
+#' @examples
+#' # make sure to prepare the plot region first
+#' jamba::nullPlot(plotAreaTitle="");
+#'
+#' # determine coordinates
+#' presetV <- c("top","bottom","left","right", "topleft");
+#' cp1 <- coordPresets(preset=presetV);
+#' cp1;
+#'
+#' points(cp1$x, cp1$y, pch=20, cex=2, col="red");
+#'
+#' # unfortunately graphics::text() does not have vectorized adj
+#' for (i in seq_along(presetV)) {
+#'    text(cp1$x[i], cp1$y[i],
+#'       labels=presetV[i],
+#'       adj=c(cp1$adjX[i], cp1$adjY[i]));
+#' }
+#'
+#' # drawLabels() is vectorized for example
+#' jamba::nullPlot(plotAreaTitle="");
+#' presetV2 <- c("topleft", "topright", "bottomleft", "bottomright");
+#' cp2 <- coordPresets(preset="center", adjPreset=presetV2, adjOffsetX=0.1, adjOffsetY=0.4);
+#' points(cp2$x, cp2$y, pch=20, cex=2, col="red");
+#' drawLabels(x=cp2$x, y=cp2$y, adjX=cp2$adjX, adjY=cp2$adjY, txt=presetV2, boxCexAdjust=c(1.15,1.6), labelCex=1.3, lx=rep(1.5, 4), ly=rep(1.5, 4))
+#'
+#' @export
+coordPresets <- function
+(preset="default",
+ x=0,
+ y=0,
+ adjPreset="default",
+ adjX=0.5,
+ adjY=0.5,
+ adjOffsetX=0,
+ adjOffsetY=0,
+ verbose=FALSE,
+ ...)
+{
+   ## Takes vector of preset values, and x,y
+   ## and re-defines x,y for anything with
+   ## top, left, right, bottom, center
+   ## Takes vector of adjPreset values, and adjX, adjY
+   ## and re-defines adjX,adjY.
+   ## When adjPreset="default" it uses opposite orientation
+   ## of "preset", otherwise position is defined by adjPreset
+   ## for top, bottom, left, right, center
+   n <- max(lengths(list(
+      x,
+      y,
+      adjX,
+      adjY,
+      preset,
+      adjPreset
+   )));
+   x <- rep(x, length.out=n);
+   y <- rep(y, length.out=n);
+   preset <- rep(preset, length.out=n);
+   adjPreset <- rep(adjPreset, length.out=n);
+   adjX <- rep(adjX, length.out=n);
+   adjY <- rep(adjY, length.out=n);
+   if (verbose) {
+      jamba::printDebug("coordPresets(): ",
+         "n:",
+         n);
+      jamba::printDebug("coordPresets(): ",
+         "preset:",
+         preset);
+      jamba::printDebug("coordPresets(): ",
+         "adjPreset:",
+         adjPreset);
+   }
+
+   ## Process the preset logic
+   x <- ifelse(grepl("right", preset), par("usr")[2],
+      ifelse(grepl("left", preset), par("usr")[1],
+         ifelse(grepl("center|top|bottom", preset), mean(par("usr")[1:2]),
+            x)));
+   if (verbose) {
+      jamba::printDebug("coordPresets(): ",
+         "x:",
+         x);
+   }
+   adjX <- ifelse(grepl("default", adjPreset),
+      ifelse(grepl("left", preset), 0-adjOffsetX,
+         ifelse(grepl("right", preset), 1+adjOffsetX,
+            ifelse(grepl("center|top|bottom", preset), 0.5,
+               adjX))),
+      ifelse(grepl("right", adjPreset), 0-adjOffsetX,
+         ifelse(grepl("left", adjPreset), 1+adjOffsetX,
+            ifelse(grepl("center|top|bottom", adjPreset), 0.5,
+               adjX))));
+   y <- ifelse(grepl("top", preset), par("usr")[4],
+      ifelse(grepl("bottom", preset), par("usr")[3],
+         ifelse(grepl("center|left|right", preset), mean(par("usr")[3:4]),
+            y)));
+   adjY <- ifelse(grepl("default", adjPreset),
+      ifelse(grepl("top", preset), 1+adjOffsetY,
+         ifelse(grepl("bottom", preset), 0-adjOffsetY,
+            ifelse(grepl("center|left|right", preset), 0.5,
+               adjY))),
+      ifelse(grepl("top", adjPreset), 0-adjOffsetY,
+         ifelse(grepl("bottom", adjPreset), 1+adjOffsetY,
+            ifelse(grepl("center|left|right", adjPreset), 0.5,
+               adjY))));
+   return(data.frame(x=x,
+      y=y,
+      adjX=adjX,
+      adjY=adjY,
+      preset=preset,
+      adjPreset=adjPreset));
+}